chore: Fix linters
